requirements.txt
pytest==8.3.3
PyYAML==6.0.2
rclone_python==0.1.12